requirements.txt
python-dotenv>=1.0.0
PyJWT>=2.9.0
pytest>=8.3.0
